requirements.txt
# Core dependencies
streamlit>=1.26.0
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.0.292
langchain-community>=0.0.10
numpy>=1.24.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0

# Document processing
pymupdf>=1.22.0
reportlab>=4.0.0
pydub>=0.25.1
trafilatura>=1.6.0

# Arabic text support
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# Audio processing
openai-whisper>=1.1.10
tensorflow>=2.13.0
tensorflow-hub>=0.14.0

# For vector embeddings
sentence-transformers>=2.2.2